+++ config.c
@@ -2255,6 +2255,15 @@ parse_section_tweak(
             LOG_WARN("tweak: grapheme shaping");
     }
 
+    else if (strcmp(key, "grapheme-width-method") == 0) {
+        if (strcmp(value, "double-width") == 0)
+            conf->tweak.grapheme_width_method = GRAPHEME_WIDTH_DOUBLE;
+        else if (strcmp(value, "wcswidth") == 0)
+            conf->tweak.grapheme_width_method = GRAPHEME_WIDTH_WCSWIDTH;
+
+        LOG_WARN("%s:%d [tweak]: grapheme-width-method=%s", path, lineno, value);
+    }
+
     else if (strcmp(key, "render-timer") == 0) {
         if (strcmp(value, "none") == 0) {
             conf->tweak.render_timer_osd = false;
@@ -2823,6 +2832,7 @@ config_load(struct config *conf, const char *conf_path,
             .fcft_filter = FCFT_SCALING_FILTER_LANCZOS3,
             .allow_overflowing_double_width_glyphs = true,
             .grapheme_shaping = false,
+            .grapheme_width_method = GRAPHEME_WIDTH_DOUBLE,
             .delayed_render_lower_ns = 500000,         /* 0.5ms */
             .delayed_render_upper_ns = 16666666 / 2,   /* half a frame period (60Hz) */
             .max_shm_pool_size = 512 * 1024 * 1024,